requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-APScheduler==1.13.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
itsdangerous==2.1.2
WTForms==3.0.1

# Database
SQLAlchemy==1.4.41

# SSH and PostgreSQL management
paramiko==3.4.1

# Web server
gunicorn==21.2.0

# Security
Werkzeug==2.3.7
bcrypt==4.0.1

# AWS SDK
boto3==1.39.3

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0